+++ software_stm32/include/compile_time.h
@@ -79,3 +79,4 @@
  
 #endif /* COMPILE_TIME_H_ */
 
+